Added Weapon Controller

+++ src/com/complex/entity/bullet/LaserBullet.java
@@ -2,7 +2,7 @@ package com.complex.entity.bullet;
 
 public class LaserBullet extends Bullet {
 
-    public LaserBullet(double x, double y, float angle, float speed) {
+    public LaserBullet(double x, double y, double angle, float speed) {
         super(x, y, angle, speed);
     }
 }
